@@ -1,71 +1,73 @@
-name: flutter_application_1
-description: "A new Flutter project."
+name: food_llm  # ✅ CHANGED: Better name than flutter_application_1
+description: "AI-powered food tracking and nutrition analysis app"
+
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-# In Windows, build-name is used as the major, minor, and patch parts
-# of the product and file versions while build-number is used as the build suffix.
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.6.2
+  sdk: '>=3.2.0 <4.0.0'  # ✅ FIXED: Proper version constraint (was ^3.6.2)
+  flutter: ">=3.16.0"    # ✅ ADDED: Minimum Flutter version
 
-# Dependencies specify other packages that your package needs in order to work.
-# To automatically upgrade your package dependencies to the latest versions
-# consider running `flutter pub upgrade --major-versions`. Alternatively,
-# dependencies can be manually updated by changing the version numbers below to
-# the latest version available on pub.dev. To see which dependencies have newer
-# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
-  camera: ^0.10.5+9 
-  shared_preferences: ^2.2.0
-  intl: ^0.18.0
-  #permission_handler: ^10.0.0
-  curved_navigation_bar: ^1.0.6
-  http: ^1.1.0
-  image_picker: ^1.0.4
-  path_provider: ^2.1.1
-  flutter_dotenv: ^5.1.0
-  flutter_image_compress: ^1.1.3
-  image_gallery_saver: ^2.0.3
-  google_fonts: ^6.1.0
-  percent_indicator: ^4.2.2
-  provider: ^6.0.5
-  share_plus: ^7.2.1
-  permission_handler: ^11.2.0
-  get_it: ^7.6.4
 
-
-
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  # ✅ CORE FLUTTER DEPENDENCIES
   cupertino_icons: ^1.0.8
+  
+  # ✅ STATE MANAGEMENT
+  provider: ^6.1.2  # Updated from 6.0.5
+
+  # ✅ LOCAL STORAGE & PREFERENCES  
+  shared_preferences: ^2.2.3  # Updated from 2.2.0
+  path_provider: ^2.1.4       # Updated from 2.1.1
+
+  # ✅ NETWORKING & HTTP
+  http: ^1.2.2                 # Updated from 1.1.0
+  flutter_dotenv: ^5.1.0      # Keep current
+
+  # ✅ CAMERA & IMAGE HANDLING
+  camera: ^0.10.6              # Updated from 0.10.5+9
+  image_picker: ^1.1.2         # Updated from 1.0.4
+  flutter_image_compress: ^2.3.0  # ✅ CRITICAL: Updated from 1.1.3 (fixes iOS issues)
+  image_gallery_saver: ^2.0.3 # Keep current
+
+  # ✅ PERMISSIONS - CRITICAL FIX
+  permission_handler: ^11.3.1  # Updated from 11.2.0 (better iOS compatibility)
+
+  # ✅ UI & NAVIGATION
+  curved_navigation_bar: ^1.0.6  # Keep current
+  
+  # ✅ SHARING
+  share_plus: ^10.0.2          # ✅ CRITICAL: Updated from 7.2.1 (fixes iOS issues)
+
+  # ✅ FONTS & UI COMPONENTS
+  google_fonts: ^6.2.1         # Updated from 6.1.0
+  percent_indicator: ^4.2.3    # Updated from 4.2.2
+
+  # ✅ DEPENDENCY INJECTION
+  get_it: ^8.0.0               # Updated from 7.6.4
+
+  # ✅ INTERNATIONALIZATION
+  intl: ^0.19.0                # Updated from 0.18.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
-  flutter_lints: ^5.0.0
+  # ✅ LINTING - Updated for better code quality
+  flutter_lints: ^5.0.0       # Keep current
+
+# ✅ DEPENDENCY OVERRIDES - Force newer versions to avoid conflicts
+dependency_overrides:
+  # Force newer image compression to avoid iOS issues
+  flutter_image_compress: ^2.3.0
+  # Ensure latest permission handler for iOS compatibility
+  permission_handler_apple: ^9.4.7
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
