@@ -1,9 +1,11 @@
+// lib/screens/splash_screen.dart
 import 'package:flutter/material.dart';
 import '../config/design_system/theme.dart';
 import '../main.dart'; // Import for MainApp placed at the top
 
 class SplashScreen extends StatefulWidget {
-  const SplashScreen({Key? key}) : super(key: key);
+  const SplashScreen({super.key});  // ✅ FIXED: Using super parameter
+
   @override
   State<SplashScreen> createState() => _SplashScreenState();
 }
@@ -40,10 +42,17 @@ class _SplashScreenState extends State<SplashScreen>
     _controller.forward();
 
     // Navigate to main app after delay
+    _navigateToMainApp();
+  }
+
+  // ✅ FIXED: Separate method with proper mounted check for BuildContext usage
+  void _navigateToMainApp() {
     Future.delayed(const Duration(seconds: 3), () {
-      Navigator.of(context).pushReplacement(
-        MaterialPageRoute(builder: (context) => const MainApp()),
-      );
+      if (mounted) {  // ✅ Guard with mounted check
+        Navigator.of(context).pushReplacement(
+          MaterialPageRoute(builder: (context) => const MainApp()),
+        );
+      }
     });
   }
 
@@ -77,7 +86,7 @@ class _SplashScreenState extends State<SplashScreen>
                         shape: BoxShape.circle,
                         boxShadow: [
                           BoxShadow(
-                            color: AppTheme.primaryBlue.withOpacity(0.3),
+                            color: AppTheme.primaryBlue.withValues(alpha: 0.3),  // ✅ FIXED: Using withValues instead of withOpacity
                             blurRadius: 20,
                             spreadRadius: 5,
                             offset: const Offset(0, 10),
